Added solana and refactor
